+++ attributes/default.rb
@@ -16,9 +16,10 @@ default['provisioner']['aws']['access_key'] = aws_config['access_key']
 default['provisioner']['aws']['region'] = aws_config['aws_region']
 
 # Knife options
-default['knife']['validation_client'] = 'test-validator'
-default['knife']['validation_key'] = '/root/.chef/test-validator.pem'
-default['provisioner']['knife']['ssl_verify'] = 'verify_none'
+# can be uncommented in the case that you are using chef client 12.0.x
+#default['knife']['validation_client'] = 'test-validator'
+#default['knife']['validation_key'] = '/root/.chef/test-validator.pem
+default['provisioner']['knife']['ssl_verify'] = ':verify_none'
 
 # Driver Configuration
 default['provisioner']['driver'] = \
@@ -29,15 +30,19 @@ default['provisioner']['driver'] = \
       'require' => 'chef/provisioning/aws_driver'
     }
   ],
-  'with-parameter' => 'aws::us-west-2'
+  'with-parameter' => 'aws::us-east-1'
 }
 default['provisioner']['driver']['machine_options'] = \
 {
-  'ssh_username' => 'root',
+  'ssh_username' => 'ubuntu',
   'use_private_ip_for_ssh' => false,
   'bootstrap_options' => {
-    'key_name' => 'dev-provisioner',
-    'image_id' => 'ami-ecb68a84',
-    'instance_type' => 'm3.medium'
+        'key_name' => 'dev-provisioner',
+        'image_id' => 'ami-ecb68a84',
+        'instance_type' => 'm3.medium'
+  },
+  'sudo' => true,
+  'convergence_options' => {
+    'ssl_verify_mode' => 'verify_none'
   }
 }